PortHamiltonianSystems.PortHamiltonianStateSpace

ControlSystemsBase.gram

ControlSystemsBase.grampd

ControlSystemsBase.ss

LinearAlgebra.norm

PortHamiltonianSystems.compose

PortHamiltonianSystems.decompose

PortHamiltonianSystems.ispassive

PortHamiltonianSystems.ispsd

PortHamiltonianSystems.kyp

PortHamiltonianSystems.kypmat

PortHamiltonianSystems.kypmax

PortHamiltonianSystems.kypmin

PortHamiltonianSystems.lrcholesky

PortHamiltonianSystems.phminreal

PortHamiltonianSystems.phss

PortHamiltonianSystems.phss

PortHamiltonianSystems.popov

PortHamiltonianSystems.prare

PortHamiltonianSystems.prgram

PortHamiltonianSystems.prgrampd

PortHamiltonianSystems.project_psd

PortHamiltonianSystems.sampopov

PortHamiltonianSystems.truncation

PortHamiltonianSystems.tsvd

PortHamiltonianSystems.unvec

PortHamiltonianSystems.unvech

PortHamiltonianSystems.PortHamiltonianStateSpace — Type

PortHamiltonianStateSpace{T}

An object representing a port-Hamiltonian state-space system.

dx(t)/dt = (J-R)Qx(t) + (G-P)u(t)
y(t)     = (G+P)'Qx(t) + (S-N)u(t)

See the function phss for a user facing constructor.

Fields:

J::SkewHermitian{T}

R::Hermitian{T}

Q::Hermitian{T}

G::Matrix{T}

P::Matrix{T}

S::Hermitian{T}

N::SkewHermitian{T}

ControlSystemsBase.gram — Method

X = gram(Σph, opt; kwargs...)

Returns the Gramian of the system Σph. If opt is :c or :o it returns the controllability or observability Gramian, respectively by calling the gram from ControlSystems package. If opt is :prc or :pro it returns the positive-real controllability or observability Gramian, respectively (see prgram).

ControlSystemsBase.grampd — Method

L = grampd(Σph, opt; kwargs...)

Returns a Cholesky factor L of the Gramian of the system Σph. If opt is :c or :o it returns the controllability or observability Gramian, respectively by calling the grampd from ControlSystems.jl package. If opt is :prc or :pro it returns a Cholesky factor L of the positive-real controllability or observability Gramian, respectively.

ControlSystemsBase.ss — Method

Σ = ss(J, R, Q, G, P, S, N)
Σ = ss(Σph)

Converts a PortHamiltonianStateSpace to a standard StateSpace.

LinearAlgebra.norm — Function

norm(Σph, p=2; kwargs...)

Converts a PortHamiltonianStateSpace to a StateSpace and calls norm on it. For more details see ControlSystems.jl package.

PortHamiltonianSystems.compose — Method

A, B, C, D = compose(J, R, Q, G, P, S, N)

Composes the port-Hamiltonian matrices to standard state-space matrices according to

A = (J - R) * Q
B = G - P
C = (G + P)' * Q
D = S - N.

PortHamiltonianSystems.decompose — Method

J, R, Q, G, P, S, N = decompose(A, B, C, D, X)

Decomposes the standard state-space matrices to port-Hamiltonian matrices for a given Hamiltonian X according to

Q = X
J = skew(A / X)
R = -sym(A / X)
G = 0.5 * (X \ C' + B)
P = 0.5 * (X \ C' - B)
S = sym(D)
N = skew(D).

PortHamiltonianSystems.ispassive — Method

ispassive(Σ::StateSpace; opt=:lmi, kwargs...)
ispassive(Σ::PortHamiltonianStateSpace; kwargs...)

Checks whether the system Σ is passive by solving the KYP inequality using kyp (if opt=:lmi) or checking the Popov function for passivity violations via sampopov (if opt=:popov).

PortHamiltonianSystems.ispsd — Method

ispsd(M)

Returns true if M is positive semi-definite, otherwise false.

PortHamiltonianSystems.kyp — Method

X = kyp(Σ; kwargs...)

Tries to solve the KYP inequality via semi definite programming.

PortHamiltonianSystems.kypmat — Method

W = kypmat(Σ, X)

Returns the KYP matrix of the system Σ for a given matrix X.

PortHamiltonianSystems.kypmax — Method

Xmin = kypmax(Σ; kwargs...)

Returns the maximal solution to the KYP inequality by solving the ARE equation for the anti-stabilizing solution.

PortHamiltonianSystems.kypmin — Method

Xmin = kypmin(Σ; kwargs...)

Returns the minimal solution to the KYP inequality by solving the ARE for the stabilizing solution.

PortHamiltonianSystems.lrcholesky — Method

L = lrcholesky(X; trunc_tol=1e-12)

Computes a low-rank approximate Cholesky-like factorization of a symmetric positive semi-definite matrix $X$ s.t. $X = L * L'$ (up to a prescribed tolerance trunc_tol).

PortHamiltonianSystems.phminreal — Method

phminreal(Σph::PortHamiltonianStateSpace; trunc_tol=1e-12)

Computes a structure preserving minimal realization of a port-Hamiltonian system [1].

PortHamiltonianSystems.phss — Method

Σph = phss(J, R, Q, G, P, S, N)
Σph = phss(J, R, Q, G)
Σph = phss(Γ, W, Q)

Creates a port-Hamiltonian state-space model Σph::PortHamiltonianStateSpace{T} with matrix element type T.

PortHamiltonianSystems.phss — Method

Σph = phss(Σ)
Σph = phss(Σ, X)

Converts a passive StateSpace to a PortHamiltonianStateSpace by executing decompose. If X is not provided, the minimal solution of the KYP inequality is used, which is obtained by solving the ARE.

PortHamiltonianSystems.popov — Method

popov(Σ, s)
popov(Σ, ω)

Evaluates the popov function of the system Σ at the complex variable s.

Φ(s) = Σ(s) + Σ(-s)'

where Σ(s) is the frequency response (transfer function) of Σ at the complex variable s.

PortHamiltonianSystems.prare — Method

prare(Σ, opt, X; kwargs...)

Evaluates the positive-real algebraic Riccati equation for system Σ and candidat solution X.

If opt is :o the positive-real controllability algebraic Riccati equation is evaluated,

A'X + XA + (C' - XB) inv(D + D') (C - B'X).

If opt is :c the positive-real observability algebraic Riccati equation is evaluated,

AX + XA' + (B - XC') inv(D + D') (B' - CX).

PortHamiltonianSystems.prgram — Method

X = prgram(Σ, opt; kwargs...)

Returns the positive-real Gramian of system Σ. If opt is :c or :o it returns the positive-real controllability or positive-real observability Gramian, respectively, by solving the corresponding positive-real algebraic Riccati equation (see prare).

PortHamiltonianSystems.prgrampd — Method

L = prgrampd(Σ, opt; kwargs...)

Returns the Cholesky factor of the positive-real Gramian of system Σ (see prgram). If opt is :c or :o it returns the positive-real controllability or positive-real observability Gramian, respectively.

In the case that the solution of the positive-real algebraic Riccati equation is not positive definite (due to numerical errors), it is projected to the set of positive semi-definite matrices calling project_psd.

PortHamiltonianSystems.project_psd — Method

Mpsd = project_psd(M; eigtol=1e-8)

Returns the nearest positive semi-definite matrix to M by setting negative eigenvalues to eigtol.

PortHamiltonianSystems.sampopov — Method

sampopov(Σ; ω=10 .^ range(-15, stop=5, length=5000))

Samples the Popov function for the system Σ at ranges of frequencies ω.

PortHamiltonianSystems.truncation — Method

truncation(d, L, trunc_tol) -> (dr, Lr)

Computes a rank revealing factorization for a given LDL-decomposition of $S = L * \mathrm{diag}(d) * L^T$ (up to a prescribed tolerance trunc_tol) such that $L_r * diag(d_r) * L_r^T \approx S$.

PortHamiltonianSystems.tsvd — Method

F = tsvd(A; kwargs...)

Returns the truncated singular value decomposition of A by truncating small singular values below ε.

PortHamiltonianSystems.unvec — Method

M = unvec(v)

Returns the matrix M from the vectorized v, i.e., the inverse of LinearAlgebra.vec.

PortHamiltonianSystems.unvech — Method

M = unvech(v)

Returns the Hermitian matrix M from the half-vectorized v, i.e., the inverse of VectorizationTransforms.vech.